requirements.txt
cloudinary==1.29.0
fastapi==0.75.2
numpy==1.21.5
opencv_python==4.5.5.62
pydantic==1.9.0
python-decouple==3.6
